ripiotest/requirements.txt
billiard==3.5.0.4
Django==1.11.14
djangorestframework==3.8.2
kombu==4.2.1
MySQL-python==1.2.5
pytz==2018.5
six==1.11.0
vine==1.1.4
